app/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
ccxt==1.54.78
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
fastapi==0.68.0
h11==0.12.0
idna==3.2
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
pycares==4.0.0
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
yarl==1.6.3
